Order CRUD , ApiClient refactoring

+++ WMS FE ASP NET Core Web/WMS FE ASP NET Core Web/Controllers/CustomerController.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authentication.Cookies;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using WMS_FE_ASP_NET_Core_Web.DTO;
 using WMS_FE_ASP_NET_Core_Web.Models;
@@ -20,13 +22,26 @@ namespace WMS_FE_ASP_NET_Core_Web.Controllers
         }
 
         // GET: Customer/Details/{id}
+        [Authorize(AuthenticationSchemes = CookieAuthenticationDefaults.AuthenticationScheme)]
         public async Task<IActionResult> Details(int id)
         {
-            var customer = await _wmsApiService.GetWMSDataAsync<CustomerModel>($"/GetCustomerBy/{id}");
-            return View(customer);
+            try
+            {
+                _wmsApiService.SetAPIParams(User.Claims);
+                if (_wmsApiService.IsTokenExpired()) return RedirectToAction("Logout", "Home");
+
+                var customer = await _wmsApiService.GetWMSDataAsync<CustomerModel>($"/GetCustomerBy/{id}");
+                return View(customer);
+            }
+            catch (Exception e)
+            {
+                _logger.LogError($"Could not get customer Id={id} details. Exception Error: {e.Message}");
+                return View();
+            }
         }
 
         // GET: Customer/Create
+        [Authorize(AuthenticationSchemes = CookieAuthenticationDefaults.AuthenticationScheme)]
         public ActionResult Create()
         {
             return View();
@@ -35,11 +50,16 @@ namespace WMS_FE_ASP_NET_Core_Web.Controllers
         // POST: Customer/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(AuthenticationSchemes = CookieAuthenticationDefaults.AuthenticationScheme)]
         public async Task<IActionResult> Create(IFormCollection collection)
         {
             try
             {
+                _wmsApiService.SetAPIParams(User.Claims);
+                if (_wmsApiService.IsTokenExpired()) return RedirectToAction("Logout", "Home");
+
                 var customer = _wrapper.Bind(collection);
+
                 var newCustomer = await _wmsApiService.PostWMSDataAsync<CreateCustomerModel>(customer, $"/CreateNewCustomer");
 
                 if (newCustomer)
@@ -53,27 +73,43 @@ namespace WMS_FE_ASP_NET_Core_Web.Controllers
                     return View();
                 }
             }
-            catch
+            catch (Exception e)
             {
+                _logger.LogError($"Could not create customer. Exception Error: {e.Message}");
                 return View();
             }
         }
 
         // GET: Customer/Edit/5
+        [Authorize(AuthenticationSchemes = CookieAuthenticationDefaults.AuthenticationScheme)]
         public async Task<IActionResult> Edit(int id)
         {
-            var customer = await _wmsApiService.GetWMSDataAsync<CustomerModel>($"/GetCustomerBy/{id}");
+            try
+            {
+                _wmsApiService.SetAPIParams(User.Claims);
+                if (_wmsApiService.IsTokenExpired()) return RedirectToAction("Logout", "Home");
 
-            return View(customer);
+                var customer = await _wmsApiService.GetWMSDataAsync<CustomerModel>($"/GetCustomerBy/{id}");
+                return View(customer);
+            }
+            catch (Exception e)
+            {
+                _logger.LogError($"Could not get customer Id={id} details. Exception Error: {e.Message}");
+                return View();
+            }
         }
 
         // POST: Customer/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(AuthenticationSchemes = CookieAuthenticationDefaults.AuthenticationScheme)]
         public async Task<IActionResult> Edit(int id, IFormCollection collection)
         {
             try
-            {
+            {                
+                _wmsApiService.SetAPIParams(User.Claims);
+                if (_wmsApiService.IsTokenExpired()) return RedirectToAction("Logout", "Home");
+
                 var updatedCustomer = _wrapper.BindToUpdateCustomer(collection);
                 var updated = await _wmsApiService.UpdateWMSDataAsync<UpdateCustomerModel>(updatedCustomer, $"/Update/Customer/{id}");
                 if (updated)
@@ -86,26 +122,39 @@ namespace WMS_FE_ASP_NET_Core_Web.Controllers
                     return View();
                 }               
             }
-            catch
+            catch (Exception e)
             {
+                _logger.LogError($"Could not update customer ID={id}. Exception Error: {e.Message}");
                 return View();
             }
         }
 
         // Delete: Customer/Delete/5
+        [Authorize(AuthenticationSchemes = CookieAuthenticationDefaults.AuthenticationScheme)]
         public async Task<IActionResult> Delete(int id)
         {
-            var deleted = await _wmsApiService.DeleteWMSDataAsync($"/Delete/Customer/{id}");
-            if (deleted)
+            try
             {
-                return RedirectToAction("Customers", "Home");
-            }
-            else
-            {
-                ViewData["ErrorMessage"] = "Could not delete customer. Please try again.";
-                return View();
-            }           
-        }
+                _wmsApiService.SetAPIParams(User.Claims);
+                if (_wmsApiService.IsTokenExpired()) return RedirectToAction("Logout", "Home");
 
+                var deleted = await _wmsApiService.DeleteWMSDataAsync($"/Delete/Customer/{id}");
+                if (deleted)
+                {
+                    return RedirectToAction("Customers", "Home");
+                }
+                else
+                {
+                    ViewData["ErrorMessage"] = "Could not delete customer. Please try again.";
+                    return View();
+                }
+
+            }
+            catch (Exception e)
+            {
+                _logger.LogError($"Could not delete customer ID={id}. Exception Error: {e.Message}");
+                return View();
+            }
+        }
     }
 }
